[selected-network-controller] Move test messenger-building functions into individual test files
