--- packages/selected-network-controller/tests/utils.ts
@@ -1,42 +1,0 @@
-import { ControllerMessenger } from '@metamask/base-controller';
-import type { NetworkControllerStateChangeEvent } from '@metamask/network-controller';
-
-import type {
-  SelectedNetworkControllerAction,
-  SelectedNetworkControllerEvent,
-} from '../src/SelectedNetworkController';
-
-/**
- * Build a controller messenger that includes all events used by the selected network
- * controller.
- *
- * @returns The controller messenger.
- */
-export function buildMessenger() {
-  return new ControllerMessenger<
-    SelectedNetworkControllerAction,
-    SelectedNetworkControllerEvent | NetworkControllerStateChangeEvent
-  >();
-}
-/**
- * Build a restricted controller messenger for the selected network controller.
- *
- * @param messenger - A controller messenger.
- * @returns The network controller restricted messenger.
- */
-export function buildSelectedNetworkControllerMessenger(
-  messenger = buildMessenger(),
-) {
-  messenger.registerActionHandler(
-    'NetworkController:getNetworkClientById',
-    jest.fn().mockReturnValue({
-      provider: { sendAsync: jest.fn() },
-      blockTracker: { getLatestBlock: jest.fn() },
-    }),
-  );
-  return messenger.getRestricted({
-    name: 'SelectedNetworkController',
-    allowedActions: ['NetworkController:getNetworkClientById'],
-    allowedEvents: ['NetworkController:stateChange'],
-  });
-}
